Tutorial fix (#123)
ramezquitao
Member
ramezquitao commented now

As reported in #105 the tutorial file was not running. This was fixed, and even thought the Tutorial.1.ipynb is far from being complete or well explained it now runs. Most of the examples were already running in the readthedocs documentation. Some time ago, an aliases "library" was implemented, so the short names for materials could be defined there (as an alias to the "real" material), without affecting the refractiveindex.info database. For the moment, adding a deprecation warning (as requested in #105) is not easy to implement. A new material class "Tabulated_N" was added to include some materials defined in the database that do not have an equation defined.

Fix #105