An option to limit the propagation in resonant cavities was added (#163)
pyOpTools failed to simulate resonand cavities. This was solved by adding 2 limits to the possible number of iterations in the propagation of a ray. One is to limit the maximum number o parents, a ray can have, another is the intensity threshold for a ray to be propagated. This 2 options are provided to the System.

solves #158