Fix to #173 (#174)
A fix to #173 was added. Also a bug in Aspherical class was found and corrected. The initial conditions on the iterative algorithm to find the ray intersection were wrong. A bug in the PowellLens component was corrected (it was using poly2d andd not the renamed Poly2D). The nofilter name was fixed in Surface.

fixes #173